% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts.R
\name{arima.conf}
\alias{arima.conf}
\title{Confidence Functions for the Coefficients of an ARIMA Model}
\usage{
arima.conf(mod)
}
\arguments{
\item{mod}{an output from stats::arima or astsa::sarima}
}
\value{
A list of lists containing the confidence functions pconf, dconf, cconf, and qconf
        for each coefficient of the ARIMA model, as well as the P-curve and S-curve.
}
\description{
Approximate confidence functions for the coefficients of an ARIMA model under
Gaussian noise assumptions.
}
\examples{
# Prediction of luteinizing hormone time series.

data(lh)

mod <- arima(lh, order = c(3,0,0))

beta.conf <- arima.conf(mod)

plot.cconf(beta.conf$ar1)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
Peter J. Brockwell and Richard A. Davis. Introduction to Time Series and Forecasting. Springer, New York, 1996. Sections 3.3 and 8.3.
}
