% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{bootstrap.beran.conf}
\alias{bootstrap.beran.conf}
\title{Simultaneous Confidence Functions via Bootstrapping.}
\usage{
bootstrap.beran.conf(bc, which = NULL)
}
\arguments{
\item{bc}{an object returned by either bcaboot or percboot}

\item{which}{the indices for the parameters to include}

\item{plot}{whether to plot the confidence density and curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for each parameter with balanced correction for multiple comparisons.
}
\description{
Confidence functions via bootstrapping based on the
percentile confidence interval, with balancing across
parameters with Beran's method for simultaneous confidence sets.
}
\examples{
# Lorem ipsum.

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, Likelihood, Probability. Vol. 41. Cambridge University Press, 2016.

             Tore Schweder. "Confidence nets for curves." Advances In Statistical Modeling And Inference: Essays in Honor of Kjell A Doksum. 2007. 593-609.

             Rudolf Beran. "Balanced simultaneous confidence sets." Journal of the American Statistical Association 83.403 (1988): 679-686.
}
