% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{chisq.conf}
\alias{chisq.conf}
\title{Confidence Functions for Pearson's Chi-squared Parameter}
\usage{
chisq.conf(N, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{N}{an I x J matrix of counts}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for Pearson's chi-squared parameter, as well as
        the P-curve and S-curve.
}
\description{
Confidence functions for Pearson's chi-squared parameter for an I x J
contingency table using the profile likelihood.
}
\examples{
# Galton's data on assortative mating according to temper.

# Columns: wife (good / bad) temper
# Rows:    husband (good / bad) temper

# reported on page 78 of *Confidence, Likelihood, Probability*.

N <- matrix(c(24, 27,
             34, 26), nrow = 2, byrow = TRUE)

chisq.conf(N)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
