% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{confdens.perc}
\alias{confdens.perc}
\title{Construct the Confidence Density from a Percentile Bootstrap for a Given Statistic}
\usage{
confdens.perc(bc, param)
}
\arguments{
\item{bc}{an object returned by percboot}

\item{param}{which parameter value to construct the percentile confidence density for}
}
\value{
A spline approximation the percentile confidence density.
}
\description{
Construct the percentile confidence density from a bootstrap sample constructed using
percboot.
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
