% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{confdens}
\alias{confdens}
\title{Construct the Confidence Density from a BCa Bootstrap for a Given Statistic}
\usage{
confdens(bc, param)
}
\arguments{
\item{bc}{an object returned by bcaboot}

\item{param}{which parameter value to construct the BCa confidence density for}
}
\value{
A spline approximation the BCa confidence density.
}
\description{
Construct the BCa confidence density from a bootstrap sample constructed using
bcaboot.
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
