% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{confdist.perc}
\alias{confdist.perc}
\title{Construct the Confidence Distribution from a Percentile Bootstrap for a Given Statistic}
\usage{
confdist.perc(bc, theta, param)
}
\arguments{
\item{bc}{an object returned by percboot}

\item{theta}{the parameter value at which to evaluate the percentile confidence distribution}

\item{param}{which parameter value to construct the percentile confidence distribution for}
}
\value{
The percentile confidence distribution for param evaluated at theta.
}
\description{
Construct the percentile confidence distribution from a bootstrap sample constructed using
bcaboot.
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
