% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{confdist}
\alias{confdist}
\title{Construct the Confidence Distribution from a BCa Bootstrap for a Given Statistic}
\usage{
confdist(bc, theta, param)
}
\arguments{
\item{bc}{an object returned by bcaboot}

\item{theta}{the parameter value at which to evaluate the BCa confidence distribution}

\item{param}{which parameter value to construct the BCa confidence distribution for}
}
\value{
The BCa confidence distribution for param evaluated at theta.
}
\description{
Construct the BCa confidence distribution from a bootstrap sample constructed using
bcaboot.
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
