% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{conffuns.from.percboot}
\alias{conffuns.from.percboot}
\title{Construct Confidence Functions from an Output of percboot}
\usage{
conffuns.from.percboot(bc)
}
\arguments{
\item{bc}{an output from percboot}
}
\value{
A list or list of lists containing the confidence functions pconf, dconf, cconf, and qconf
        for the parameter associated with the statistic(s) used with percboot.
}
\description{
Construct confidence functions from the output of percboot.
This is a helper function to make constructing ones own
bootstrap confidence functions straightforward.
}
\examples{
# Reproduce percentile confidence density from Figure 11.7
# of *Computer Age Statistical Inference*.

scor <-
read.table('https://web.stanford.edu/~hastie/CASI_files/DATA/student_score.txt',
header = TRUE)

statistic <- function(data, id = 1:nrow(data), ...){
  dat <- data[id, ]

  Sigma <- cov(dat)*((nrow(dat)-1)/nrow(dat))

  lams <- eigen(Sigma, symmetric = TRUE, only.values = TRUE)$values

  return(lams[1])
}

ran.gen <- function(data, mle){
  MASS::mvrnorm(n = nrow(data), mle$mu, mle$Sigma)
}

bc <- percboot(data = scor, statistic = statistic, B = 8000, sim = "parametric", ran.gen = ran.gen,
              mle = list(mu = colMeans(scor),
                         Sigma = cov(scor)*((nrow(scor)-1)/nrow(scor))))

lam.conf <- conffuns.from.percboot(bc)

plot.cconf(lam.conf, xlab = 'Largest Eigenvalue')
plot.dconf(lam.conf, xlab = 'Largest Eigenvalue')

lam.conf$qconf(c(0.025, 0.975))

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
