% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{confquant.perc}
\alias{confquant.perc}
\title{Construct the Confidence Quantile Function from a Percentile Bootstrap for a Given Statistic}
\usage{
confquant.perc(bc, p, param)
}
\arguments{
\item{bc}{an object returned by percboot}

\item{p}{the desired probability value at which to evaluate the confidence quantile function}

\item{param}{which parameter value to construct the percentile confidence quantile function for}
}
\value{
The percentile confidence quantile function for param evaluated at p.
}
\description{
Construct the percentile confidence quantile function from a bootstrap sample constructed using
bcaboot.
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
