% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{confquant}
\alias{confquant}
\title{Construct the Confidence Quantile Function from a BCa Bootstrap for a Given Statistic}
\usage{
confquant(bc, p, param)
}
\arguments{
\item{bc}{an object returned by bcaboot}

\item{p}{the desired probability value at which to evaluate the confidence quantile function}

\item{param}{which parameter value to construct the BCa confidence quantile function for}
}
\value{
The BCa confidence quantile function for param evaluated at p.
}
\description{
Construct the BCa confidence quantile function from a bootstrap sample constructed using
bcaboot.
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
