% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor.conf}
\alias{cor.conf}
\title{Confidence Functions for Pearson's Correlation Coefficient}
\usage{
cor.conf(x, y, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}

\item{"x, }{y" numeric vectors of data values. x and y must have the same length.}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for Pearson's correlation coefficient via Fisher's Z-transformation, as well as
        the P-curve and S-curve.
}
\description{
Confidence functions for Pearson's correlation coefficient
via Fisher's Z-transformation.
}
\examples{
data(fat)
cor.conf(x = fat$body.fat, y = fat$weight)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
