% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{dcorr}
\alias{dcorr}
\title{Density of the Sample Pearson Correlation for a Bivariate Gaussian}
\usage{
dcorr(x, rho, n)
}
\arguments{
\item{x}{vector of quantiles}

\item{rho}{the population correlation coefficient}

\item{n}{the sample size}
}
\value{
A vector containing the density function evaluated at x.
}
\description{
The density of the sample Pearson correlation for a bivariate Gaussian.
}
\references{
Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
