% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dietstudy}
\alias{dietstudy}
\title{Weight Loss Due to Healthy Low Carb versus Healthy Low Fat Diet}
\format{
A data frame with 609 rows and 2 columns:
\describe{
  \item{weightchange}{weight change after 12 months in pounds.}
  \item{diet}{factor indicating prescribed diet of participant}
}
}
\source{
Christopher D. Gardner, John F. Trepanowski, Liana C. Del Gobbo, Michelle E. Hauser, Joseph Rigdon, John P. A. Ioannidis, Manisha Desai, and Abby C. King. "Effect of low-fat vs low-carbohydrate diet on 12-month weight loss in overweight adults and the association with genotype pattern or insulin secretion: the DIETFITS randomized clinical trial." *JAMA* 319, no. 7 (2018): 667-679.
}
\usage{
dietstudy
}
\description{
The `dietstudy` data frame has 609 rows and 2 columns. The data are a
**synthetic** data set made to match the summary statistics from the study
}
\details{
"Effect of Low-Fat vs Low-Carbohydrate Diet on 12-Month Weight Loss in
Overweight Adults and the Association With Genotype Pattern
or Insulin Secretion"

from JAMA.

The data consist of weight changes (negative indicates weight loss)
12 months after direction to follow a Healthy Low Carb (HLC) or Healthy
Low Fat (HLF) diet.
}
\keyword{datasets}
