% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{dncd}
\alias{dncd}
\title{Probability Mass Function for Noncentral Hypergeometric Distribution}
\usage{
dncd(y2, n1, n2, z, rho)
}
\arguments{
\item{y2}{the number of successes in the second group}

\item{n1}{the sample size for the first group}

\item{n2}{the sample size for the second group}

\item{z}{the total number of successes in both groups}

\item{rho}{the odds ratio}
}
\value{
The probability of y2 successes in the second group,
        conditional on z successes overall, with an odds
        ratio of rho.
}
\description{
Computes the probability mass function for the
noncentral hypergeometric distribution.
}
