% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fat}
\alias{fat}
\title{Body Measurements for Predicting Body Fat Percentage in Males}
\format{
A data frame with 247 rows and 15 columns:
\describe{
  \item{body.fat}{estimated body fat percentage using water displacement.}
  \item{age}{age in years.}
  \item{weight}{weight in pounds.}
  \item{height}{height in inches.}
  \item{BMI}{Body Mass Index.}
  \item{neck}{neck circumference in cm.}
  \item{chest}{chest circumference in cm.}
  \item{abdomen}{abdomen circumference in cm.}
  \item{hip}{hip circumference in cm.}
  \item{thigh}{thigh circumference in cm.}
  \item{knee}{knee circumference in cm.}
  \item{ankle}{ankle circumference in cm.}
  \item{bicep}{bicep circumference in cm.}
  \item{forearm}{forearm circumference in cm.}
  \item{wrist}{wrist circumference in cm.}
}
}
\source{
This data set comes from the collection of the Journal of Statistics
Education at \url{http://www.amstat.org/publications/jse/datasets/fat.txt}.
The data set was contributed by Roger W. Johnson.
}
\usage{
fat
}
\description{
The `fat` data frame has 247 rows and 15 columns. These consist of physical
measurements of 247 men with ages ranging from 22 to 81.
}
\details{
This data set can be used to demonstrate prediction of body fat
percentage (as estimated by water displacement) via easier body measurements
like height, weight, various circumferences, etc.
}
\references{
The source of the data is attributed to Dr. A. Garth Fisher,
Human Performance Research Center, Brigham Young University,
Provo, Utah 84602,
}
\keyword{datasets}
