% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.beta.conf}
\alias{glm.beta.conf}
\title{Confidence Functions for the Coefficients of a GLM}
\usage{
glm.beta.conf(mod)
}
\arguments{
\item{mod}{an output from glm}
}
\value{
A list of lists containing the confidence functions pconf, dconf, cconf, and qconf
        for each coefficient of the glm, as well as the P-curve and S-curve.
}
\description{
Confidence functions for the coefficients of a glm.
}
\examples{
# Low birth weight example from page 38 of *Confidence, Likelihood, Probability*,

data(lbw)

mod <- glm(low ~ weight + age + black + other + smoker, data = lbw, family = binomial)

mod.conf <- glm.beta.conf(mod)

display.cconf(mod.conf$weight)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
