% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.lincom.conf.disp}
\alias{glm.lincom.conf.disp}
\title{Confidence Functions for the Expected Response of a GLM with Dispersion}
\usage{
glm.lincom.conf.disp(mod, x, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{mod}{an output from glm}

\item{x}{a vector containing the predictor values for which the expected
response is desired}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for the expected response at x, as well as the P-curve and S-curve.
}
\description{
Confidence functions for the expected response of a GLM with dispersion.
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
