% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{glmer.beta.conf}
\alias{glmer.beta.conf}
\title{Confidence Functions for the Coefficients of a LMM / GLMM}
\usage{
glmer.beta.conf(mod)
}
\arguments{
\item{mod}{an output from glmer}
}
\value{
A list of lists containing the confidence functions pconf, dconf, cconf, and qconf
        for each coefficient of the GLMM, as well as the P-curve and S-curve.
}
\description{
Confidence functions for the Coefficients of a (Generalized) Linear Mixed Model.
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
