% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lbw}
\alias{lbw}
\title{Risk Factors Associated with Low Infant Birth Weight}
\format{
A data frame with 189 rows and 6 columns:
\describe{
  \item{low}{factor for whether the birth weight is <= 2.5 kg or > 2.5 kg.}
  \item{age}{age of the mother in years.}
  \item{weight}{weight of mother in kg at last menstrual period.}
  \item{black}{indicator for mother's race being black.}
  \item{other}{indicator for mother's race being other.}
  \item{smoker}{indicator for mother's smoking status during pregnancy.}
}
}
\source{
Hosmer, D.W. and Lemeshow, S. (1989) *Applied Logistic Regression*. New York: Wiley
}
\usage{
lbw
}
\description{
The `lbw` data frame has 189 rows and 6 columns. The data were collected at
Baystate Medical Center, Springfield, Mass during 1986. The original data
set is from the `MASS` package.
}
\references{
Venables, W. N. and Ripley, B. D. (2002) *Modern Applied Statistics with S*. Fourth edition. Springer.
}
\keyword{datasets}
