% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{lm.beta.boot.conf}
\alias{lm.beta.boot.conf}
\title{Bootstrapped Confidence Functions for One or Two Means}
\usage{
lm.beta.boot.conf(formula, data, B = 2000)
}
\arguments{
\item{B}{the number of bootstrap samples used to approximate the
bootstrap distribution}

\item{x}{a vector containing the first sample}

\item{y}{a vector containing the second sample (optional)}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for a single mean or the difference of two means, as well as
        the P-curve and S-curve.
}
\description{
Bootstrapped confidence functions for a single mean or the difference
of two means using the BCa bootstrap.
}
\examples{
data(fat)

beta.boot.conf <- lm.beta.boot.conf(body.fat ~ age + weight + height, data = fat, B = 2000)

plot.cconf(beta.boot.conf$weight)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
