% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm.beta.conf}
\alias{lm.beta.conf}
\title{Confidence Functions for the Coefficients of a Linear Model}
\usage{
lm.beta.conf(mod)
}
\arguments{
\item{mod}{an output from lm}
}
\value{
A list of lists containing the confidence functions pconf, dconf, cconf, and qconf
        for each coefficient of the linear model, as well as the P-curve and S-curve.
}
\description{
Confidence functions for the coefficients of a linear model under
Gaussian noise assumptions.
}
\examples{
# Prediction of body fat percentage from other body measurements.

data(fat)

mod <- lm(body.fat ~ age + weight + height, data = fat)

beta.conf <- lm.beta.conf(mod)

plot.cconf(beta.conf$weight)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
