% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm.lincom.conf}
\alias{lm.lincom.conf}
\title{Confidence Functions for the Expected Response of a Linear Model}
\usage{
lm.lincom.conf(mod, x, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{mod}{an output from lm}

\item{x}{a vector containing the predictor values for which the expected
response is desired}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for the expected response at x, as well as the P-curve and S-curve.
}
\description{
Confidence functions for the expected response of a linear model under
Gaussian noise assumptions.
}
\examples{

# Prediction of body fat percentage from other body measurements.

data(fat)

mod <- lm(body.fat ~ age + weight + height, data = fat)

# Confidence curve for the expected BFP of a 170lb, 6ft, 20-year old male

x <- c(1,    # Intercept
       20,   # 20 years old
       170,  # 170 lbs
       6*12) # 72 inches (6 feet)

m.conf <- lm.lincom.conf(mod, x)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
