% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{lm.sigma.conf}
\alias{lm.sigma.conf}
\title{Confidence Functions for the Noise Standard Deviation of a Linear Model}
\usage{
lm.sigma.conf(mod, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{mod}{an output from lm}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for the noise standard deviation, as well as the P-curve and S-curve.
}
\description{
Confidence functions for the noise standard deviation of a linear model under
Gaussian noise assumptions.
}
\examples{
# Prediction of body fat percentage from other body measurements.

data(fat)

mod <- lm(body.fat ~ age + weight + height, data = fat)

sigma.conf <- lm.sigma.conf(mod)

plot.cconf(sigma.conf)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
