% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multinomial.conf}
\alias{multinomial.conf}
\title{Simultaneous Confidence Functions for Multinomial Proportions.}
\usage{
multinomial.conf(N, plot = TRUE, conf.level = 0.95, B = 2000, col = NULL)
}
\arguments{
\item{N}{a vector of counts in each category from a sample of size n}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}

\item{B}{the number of bootstrap samples to use with Beran's method}

\item{col}{a vector of colors to use for plotting confidence functions of each proportion}
}
\value{
A list of lists of lists containing the confidence functions pconf, dconf, cconf, and qconf
        for each proportion without correction for multiple comparison (singlecomp) or with
        correction for multiple comparison (multicomp).
}
\description{
Confidence functions for a multinomial proportions based on the
mid P-value for each proportion, with balancing across
proportions via bootstrapping with Beran's simultaneous
confidence sets.
}
\examples{
# A hypothetical Punnett square experiment with peas.

N <- c(315, 108, 101, 32)

names(N) <- c('Round, Yellow', 'Round, Green', 'Wrinkled, Yellow', 'Wrinkled, Green')

col <- c('darkgoldenrod', 'darkgreen', 'darkgoldenrod1', 'darkolivegreen')

peas.conf <- multinomial.conf(N, col = col)

# Confidence intervals without and with correction for multiple comparisons:
peas.conf$singlecomp$`Round, Yellow`$qconf(c(0.025, 0.975))
peas.conf$multicomp$`Round, Yellow`$qconf(c(0.025, 0.975))

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, Likelihood, Probability. Vol. 41. Cambridge University Press, 2016.

             Tore Schweder. "Confidence nets for curves." Advances In Statistical Modeling And Inference: Essays in Honor of Kjell A Doksum. 2007. 593-609.

             Rudolf Beran. "Balanced simultaneous confidence sets." Journal of the American Statistical Association 83.403 (1988): 679-686.
}
