% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{p.multinomial}
\alias{p.multinomial}
\title{Compute a P-value for Specified Multinomial Proportions}
\usage{
p.multinomial(obj, theta)
}
\arguments{
\item{obj}{an output from conf.multinomial().}

\item{theta}{a vector as the same length as the N used
to construct obj; the null value of the
multinomial proportions.}
}
\value{
A P-value from theta.
}
\description{
Compute a P-value for specified multinomial proportions
using the confidence net returned by conf.multinomial().
}
\examples{
# A hypothetical Punnett square experiment with peas.

N <- c(315, 108, 101, 32)

names(N) <- c('Round, Yellow', 'Round, Green', 'Wrinkled, Yellow', 'Wrinkled, Green')

col <- c('darkgoldenrod', 'darkgreen', 'darkgoldenrod1', 'darkolivegreen')

peas.conf <- multinomial.conf(N, col = col)

p.multinomial(peas.conf, c(9, 3, 3, 1)/16)

}
