% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{pcorr}
\alias{pcorr}
\title{Distribution Function of the Sample Pearson Correlation for a Bivariate Gaussian}
\usage{
pcorr(q, rho, n, lower.tail = TRUE)
}
\arguments{
\item{rho}{the population correlation coefficient}

\item{n}{the sample size}

\item{x}{vector of quantiles}
}
\value{
A vector containing the distribution function evaluated at x.
}
\description{
The distribution function of the sample Pearson correlation for a bivariate Gaussian.
}
\references{
Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
