% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{percboot}
\alias{percboot}
\title{Perform Percentile Bootstrap for a Given Statistic}
\usage{
percboot(
  data,
  statistic,
  B = 2000,
  sim = "ordinary",
  stratified = FALSE,
  ran.gen = function(d, p) d,
  mle = NULL,
  formula = NULL
)
}
\arguments{
\item{data}{the dataframe or data matrix}

\item{statistic}{a function that computes the statistic from data}

\item{B}{the number of bootstrap replicates}

\item{sim}{either "ordinary" (for case resampling bootstrap) or "parametric" (for parametric bootstrap)}

\item{stratified}{whether or not (default) to use the use stratified
sampling for the bootstrapping.}

\item{ran.gen}{a function returning a random sample, for the parametric bootstrap}

\item{mle}{the maximum likelihood estimate from the original sample, for the parametric bootstrap}

\item{formula}{for use with interfacing to lm, glm, etc.}
}
\value{
A list containing the sample estimate, bootstrap estimates,
        and bootstrap CDF for the percentile bootstrap.
}
\description{
Approximate the bootstrap distribution for the percentile bootstrap, given a statistic as would be supplied to
boot from the boot package.
}
\examples{
# Bootstrap confidence functions for a single mean.

t.one.sample <- function(data, id = 1:length(data), ...){
  dat <- data[id]

  d <- mean(dat)

  return(d)
}

data(dietstudy)

percboot(data = dietstudy$weightchange[dietstudy$diet == 'Low Carb'],
        statistic = t.one.sample,
        B = 2000)


}
\references{
Thomas J. DiCiccio and Bradley Efron. "Bootstrap confidence intervals." Statistical Science (1996): 189-212.
}
