% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cconf}
\alias{plot.cconf}
\title{A Wrapper Function for Plotting Confidence Curves}
\usage{
\method{plot}{cconf}(x, conf.level = 0.95, n.points = 2001, ...)
}
\arguments{
\item{x}{a list of confidence functions returned by a *.conf() function}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}

\item{n.points}{the number of points to plot for the confidence curve}

\item{...}{additional plotting parameters.}
}
\value{
a plot of the confidence curve from x
}
\description{
A wrapper for plotting confidence curves using confidence objects
returned by *.conf() functions from clp.
}
\examples{
data(dietstudy)

x <- dietstudy$weightchange[dietstudy$diet == 'Low Carb']

out <- t.conf(x, plot = FALSE)

plot.cconf(out, conf.level = 0.999, xlab = 'Average Weight Loss in Low Carb Group (lbs)')

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
