% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pconf}
\alias{plot.pconf}
\title{A Wrapper Function for Plotting Confidence Distributions}
\usage{
\method{plot}{pconf}(obj, xlim = NULL, xlab = NULL, col = "black", n.points = 2001)
}
\arguments{
\item{obj}{a list of confidence functions returned by a *.conf() function}

\item{xlim}{the horizontal limits for the plot. If NULL, taken as a 99.8\%
confidence interval.}

\item{xlab}{the label for the horizontal axis}

\item{col}{the color to plot the confidence curve}

\item{n.points}{the number of points to plot for the confidence curve}

\item{log}{switch for plotting the horizontal ('x') or vertical ('x') or both ('xy') axes with a
log-scaling}
}
\value{
a plot of the confidence curve from obj
}
\description{
A wrapper for plotting confidence distributions using confidence objects
returned by *.conf() functions from clp.
}
\examples{
data(dietstudy)

x <- dietstudy$weightchange[dietstudy$diet == 'Low Carb']

out <- t.conf(x, plot = FALSE)

plot.pconf(out, xlab = 'Average Weight Loss in Low Carb Group (lbs)')

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.
}
