% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.scurve}
\alias{plot.scurve}
\title{A Wrapper Function for Plotting S-Curves}
\usage{
\method{plot}{scurve}(
  obj,
  conf.level = 0.95,
  xlim = NULL,
  xlab = NULL,
  log = "",
  col = "black",
  n.points = 2001
)
}
\arguments{
\item{obj}{a list of confidence functions returned by a *.conf() function}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}

\item{xlim}{the horizontal limits for the plot. If NULL, taken as the larger of
a conf.level confidence interval or a 1 - (1 - conf.level)/2 confidence interval.}

\item{xlab}{the label for the horizontal axis}

\item{log}{switch for plotting the horizontal ('x') or vertical ('x') or both ('xy') axes with a
log-scaling}

\item{col}{the color to plot the confidence curve}

\item{n.points}{the number of points to plot for the confidence curve}
}
\value{
a plot of the S-curve from obj
}
\description{
A wrapper for plotting S-curves using confidence objects
returned by *.conf() functions from clp.
}
\examples{
data(dietstudy)

x <- dietstudy$weightchange[dietstudy$diet == 'Low Carb']

out <- t.conf(x, plot = FALSE)

plot.scurve(out, conf.level = 0.999, xlab = 'Average Weight Loss in Low Carb Group (lbs)')

}
\references{
Zad R. Chow and Sander Greenland. "Semantic and Cognitive Tools to Aid Statistical Inference: Replace Confidence and Significance by Compatibility and Surprise." arXiv preprint arXiv:1909.08579 (2019).
}
