% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop.conf.1s}
\alias{prop.conf.1s}
\title{Confidence Functions for One Proportion}
\usage{
prop.conf.1s(x, n, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{x}{the number of successes out of n trials}

\item{n}{the number of trials}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for a single proportion, as well as the P-curve and S-curve.
}
\description{
Confidence functions a single binomial proportion based on the
mid P-value.
}
\examples{
prop.conf.1s(x = 0, n = 10)
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

            Jan Hannig. "On generalized fiducial inference." Statistica Sinica (2009): 491-544.
}
