% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.R
\name{prop.conf.2s}
\alias{prop.conf.2s}
\title{Confidence Functions for the Difference of Two Proportions}
\usage{
prop.conf.2s(x, n, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{x}{a vector of counts of successes in each independent sample}

\item{n}{a vector of the number of trials in each independent sample}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for the difference of two proportions, as well as the P-curve and S-curve.
}
\description{
Confidence functions for the difference of two binomial proportions
based on the score statistic.
}
\examples{
prop.conf.2s(x = c(1, 2), n = c(10, 10))
}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

            Olli Miettinen and Markku Nurminen. "Comparative analysis of two rates." Statistics in Medicine 4.2 (1985): 213-226.

            Markku Nurminen. "Confidence intervals for the difference and ratio of two binomial proportions." Biometrics 42 (1986): 675-676.
}
