% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{signtest.conf}
\alias{signtest.conf}
\title{Confidence Functions for the Median via the Sign Test}
\usage{
signtest.conf(x, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{x}{a vector of observations}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for the median, as well as the P-curve and S-curve.
}
\description{
Confidence functions for the median of a continuous distribution based on the
sign test.
}
\examples{
data(dietstudy)

x <- dietstudy$weightchange[dietstudy$diet == 'Low Carb']
y <- dietstudy$weightchange[dietstudy$diet == 'Low Fat']

lc.signtest.conf <- signtest.conf(x)
lf.signtest.conf <- signtest.conf(y)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Myles Hollander, Douglas A. Wolfe, and Eric Chicken. Nonparametric Statistical Methods. Vol. 751. John Wiley & Sons, 2013.
}
