% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{t.boot.conf}
\alias{t.boot.conf}
\title{Bootstrapped Confidence Functions for One or Two Means}
\usage{
\method{t}{boot.conf}(x, y = NULL, B = 2000, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{x}{a vector containing the first sample}

\item{y}{a vector containing the second sample (optional)}

\item{B}{the number of bootstrap samples used to approximate the
bootstrap distribution}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for a single mean or the difference of two means, as well as
        the P-curve and S-curve.
}
\description{
Bootstrapped confidence functions for a single mean or the difference
of two means using the BCa bootstrap.
}
\examples{
data(dietstudy)

# One mean

t.boot.conf(x = dietstudy$weightchange[dietstudy$diet == 'Low Carb'],
            B = 2000)

# Two means

t.boot.conf(x = dietstudy$weightchange[dietstudy$diet == 'Low Carb'],
            y = dietstudy$weightchange[dietstudy$diet == 'Low Fat'],
            B = 2000)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
