% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{t.conf}
\alias{t.conf}
\title{Confidence functions for t-Test}
\usage{
\method{t}{conf}(x, y = NULL, paired = FALSE, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{plot}{a logical indicating whether you want the plots to be generated.}
}
\value{
A list containing the confidence functions.
}
\description{
Computes confidence functions for one and two sample t-tests on vectors of data.
}
\examples{
t.confdist(x = 1:10, y = 7:20)

}
