% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{t.one.sample}
\alias{t.one.sample}
\title{Statistic for Bootstrapped One-sample t-test.}
\usage{
\method{t}{one.sample}(data, id = 1:length(data), ...)
}
\arguments{
\item{data}{a vector containing the sample}

\item{id}{the id variable used by boot}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A single value of the bootstrapped sample mean.
}
\description{
The function to pass to bcaboot to perform the one-sample t-test
via bootstrapping.
}
