% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{wilcox.conf}
\alias{wilcox.conf}
\title{Confidence Functions for the (Pseudo)Median or Shift via Wilcoxon Tests}
\usage{
wilcox.conf(x, y = NULL, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{x}{a vector of observations for the first sample}

\item{y}{a vector of observations for the second sample, if applicable}

\item{plot}{whether to plot the confidence density and curve}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for the median, as well as the P-curve and S-curve.
}
\description{
Confidence functions for the (pseudo)median of a continuous distribution
via the Wilcoxon signed-rank test (for one sample) or shift between
two continuous distributions via the Wilcoxon rank-sum test (for two samples).
}
\examples{
data(dietstudy)

x <- dietstudy$weightchange[dietstudy$diet == 'Low Carb']
y <- dietstudy$weightchange[dietstudy$diet == 'Low Fat']

lc.wc.conf <- wilcox.conf(x)
lf.wc.conf <- wilcox.conf(y)

comp.wc.conf <- wilcox.conf(x, y)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Myles Hollander, Douglas A. Wolfe, and Eric Chicken. Nonparametric Statistical Methods. Vol. 751. John Wiley & Sons, 2013.
}
