% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{bcaboot}
\alias{bcaboot}
\title{Perform BCa Bootstrap for a Given Statistic}
\usage{
bcaboot(
  data,
  statistic,
  B = 2000,
  sim = "ordinary",
  stratified = FALSE,
  ran.gen = function(d, p) d,
  mle = NULL,
  formula = NULL
)
}
\arguments{
\item{data}{the dataframe or data matrix}

\item{statistic}{a function that computes the statistic from data}

\item{B}{the number of bootstrap replicates}

\item{sim}{either "ordinary" (for case resampling bootstrap) or "parametric" (for parametric bootstrap)}

\item{stratified}{whether or not (default) to use the use stratified
sampling for the bootstrapping.}

\item{ran.gen}{a function returning a random sample, for the parametric bootstrap}

\item{mle}{the maximum likelihood estimate from the original sample, for the parametric bootstrap}

\item{formula}{for use with interfacing to lm, glm, etc.}
}
\value{
A list containing the sample estimate, bootstrap estimates,
        bootstrap CDF, bias, and acceleration parameters for the
        BCa bootstrap.
}
\description{
Approximate the bootstrap distribution and estimate bias and acceleration
parameters for the BCa bootstrap, given a statistic as would be supplied to
boot from the boot package.
}
\examples{
# Bootstrap confidence functions for a single mean.

t.one.sample <- function(data, id = 1:length(data)){
  dat <- data[id]

  d <- mean(dat)

  return(d)
}

data(dietstudy)

bcaboot(data = dietstudy$weightchange[dietstudy$diet == 'Low Carb'],
        statistic = t.one.sample,
        B = 2000)

# Reproduce BCa confidence density from Figure 11.7
# of *Computer Age Statistical Inference*.

scor <- read.table('https://web.stanford.edu/~hastie/CASI_files/DATA/student_score.txt', header = TRUE)

statistic <- function(data, id = 1:nrow(data)){
  dat <- data[id, ]

  Sigma <- cov(dat)*((nrow(dat)-1)/nrow(dat))

  lams <- eigen(Sigma, symmetric = TRUE, only.values = TRUE)$values

  return(lams[1])
}

ran.gen <- function(data, mle){
  mvrnorm(n = nrow(data), mle$mu, mle$Sigma)
}

bc <- bcaboot(data = scor, statistic = statistic, B = 8000, sim = "parametric", ran.gen = ran.gen,
              mle = list(mu = colMeans(scor),
                         Sigma = cov(scor)*((nrow(scor)-1)/nrow(scor))))

}
\references{
Bradley Efron. "Better bootstrap confidence intervals." Journal of the American Statistical Association 82.397 (1987): 171-185.

             Thomas J. DiCiccio and Bradley Efron. "Bootstrap confidence intervals." Statistical Science (1996): 189-212.
}
