% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{bcaboot}
\alias{bcaboot}
\title{Perform BCa Bootstrap for a Given Statistic}
\usage{
bcaboot(data, statistic, B = 2000, stratified = FALSE)
}
\arguments{
\item{data}{the dataframe or data matrix}

\item{statistic}{a function that computes the statistic from data}

\item{B}{the number of bootstrap replicates}

\item{stratified}{whether or not (default) to use the use stratified
sampling for the bootstrapping.}
}
\value{
A list containing the sample estimate, bootstrap estimates,
        bootstrap CDF, bias, and acceleration parameters for the
        BCa bootstrap.
}
\description{
Approximate the bootstrap distribution and estimate bias and acceleration
parameters for the BCa bootstrap, given a statistic as would be supplied to
boot from the boot package.
}
\examples{
t.one.sample <- function(data, id = 1:length(data)){
  dat <- data[id]

  d <- mean(dat)

  return(d)
}

data(dietstudy)

bcaboot(data = dietstudy$weightchange[dietstudy$diet == 'Low Carb'],
        statistic = t.one.sample,
        B = 2000)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron. "Better bootstrap confidence intervals." Journal of the American Statistical Association 82.397 (1987): 171-185.

             Thomas J. DiCiccio and Bradley Efron. "Bootstrap confidence intervals." Statistical Science (1996): 189-212.
}
