% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{lm.beta.boot.conf}
\alias{lm.beta.boot.conf}
\title{Bootstrapped Confidence Functions for Coefficients of a Linear Model}
\usage{
lm.beta.boot.conf(formula, data, B = 2000)
}
\arguments{
\item{formula}{the model to be fitted}

\item{data}{the data frame containing the data for fitting}

\item{B}{the number of bootstrap samples used to approximate the
bootstrap distribution}
}
\value{
A list of lists containing the confidence functions pconf, dconf, cconf, and qconf
        for each coefficient of the linear model, as well as
        the P-curve and S-curve.
}
\description{
Bootstrapped confidence functions for coefficients of a linear model
using BCa and the case resampling bootstrap.
}
\examples{
data(fat)

beta.boot.conf <- lm.beta.boot.conf(body.fat ~ age + weight + height, data = fat, B = 2000)

plot.cconf(beta.boot.conf$weight)

}
\references{
Tore Schweder and Nils Lid Hjort. Confidence, likelihood, probability. Vol. 41. Cambridge University Press, 2016.

             Bradley Efron and Trevor Hastie. Computer Age Statistical Inference. Vol. 5. Cambridge University Press, 2016.
}
