% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{t.conf}
\alias{t.conf}
\title{Confidence Functions for Means via Gaussian Models from Data}
\usage{
\method{t}{conf}(x, y = NULL, paired = FALSE, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{plot}{a logical indicating whether you want the plots to be generated.}

\item{conf.level}{the confidence level for the confidence interval indicated on the confidence curve}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for the mean or difference of means, as well as the P-curve and S-curve.
}
\description{
Confidence functions for a single mean, the difference of means from matched
observations, or the difference of means from independent samples.
}
\examples{

# One mean:

t.conf(x = 1:10)

# Two means:

t.conf(x = 1:10, y = 7:20)

}
