% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{t.conf.summary}
\alias{t.conf.summary}
\title{Confidence Functions for Means via Gaussian Models from Summary Statistics}
\usage{
\method{t}{conf.summary}(mean, sd, n, plot = TRUE, conf.level = 0.95)
}
\arguments{
\item{mean}{a (non-empty) numeric vector of data values.}

\item{sd}{a (non-empty) numeric vector of data values.}

\item{n}{a (non-empty) numeric vector of data values.}

\item{plot}{a logical indicating whether you want the plots to be generated.}
}
\value{
A list containing the confidence functions pconf, dconf, cconf, and qconf
        for the mean or difference of means, as well as the P-curve and S-curve.
}
\description{
Confidence functions for a single mean, the difference of means from matched
observations, or the difference of means from independent samples.
}
\examples{

t.confdist(mean = c(3,5), sd = c(1,.5), n = c(10, 12))

}
